move fastcase builtins to their own file
Also cleanup some more things:
* No forward declaring functions in unrelated headers
* Mark interpreter local functions static
* Reorder strange includes
