+++ rir/src/common.cpp
@@ -1,7 +1,7 @@
 #include "common.h"
 #include "R/RList.h"
 #include "R/r.h"
-#include "runtime/ArgsLazyData.h"
+#include "interpreter/ArgsLazyData.h"
 #include "utils/capture_out.h"
 
 #include <cxxabi.h>
