+++ rir/src/interpreter/interp_builtins.cpp
@@ -1,0 +1,265 @@
+#include "interp_builtins.h"
+
+namespace rir {
+
+SEXP tryFastSpecialCall(const CallContext& call, Context* ctx) {
+    SLOWASSERT(call.hasStackArgs() && !call.hasNames());
+    return nullptr;
+}
+
+SEXP tryFastBuiltinCall(const CallContext& call, Context* ctx) {
+    SLOWASSERT(call.hasStackArgs() && !call.hasNames());
+
+    static constexpr size_t MAXARGS = 32;
+    std::array<SEXP, MAXARGS> args;
+    auto nargs = call.suppliedArgs;
+
+    if (nargs > MAXARGS)
+        return nullptr;
+
+    for (size_t i = 0; i < call.suppliedArgs; ++i) {
+        auto arg = call.stackArg(i);
+        if (TYPEOF(arg) == PROMSXP)
+            arg = PRVALUE(arg);
+        if (arg == R_UnboundValue || arg == R_MissingArg ||
+            ATTRIB(arg) != R_NilValue)
+            return nullptr;
+        args[i] = arg;
+    }
+
+    switch (call.callee->u.primsxp.offset) {
+    case 90: { // "c"
+        if (nargs == 0)
+            return R_NilValue;
+
+        auto type = TYPEOF(args[0]);
+        if (type != REALSXP && type != LGLSXP && type != INTSXP)
+            return nullptr;
+        long total = XLENGTH(args[0]);
+        for (size_t i = 1; i < nargs; ++i) {
+            auto thistype = TYPEOF(args[i]);
+            if (thistype != REALSXP && thistype != LGLSXP && thistype != INTSXP)
+                return nullptr;
+
+            if (thistype == INTSXP && type == LGLSXP)
+                type = INTSXP;
+
+            if (thistype == REALSXP && type != REALSXP)
+                type = REALSXP;
+
+            total += XLENGTH(args[i]);
+        }
+
+        if (total == 0)
+            return nullptr;
+
+        long pos = 0;
+        auto res = Rf_allocVector(type, total);
+        for (size_t i = 0; i < nargs; ++i) {
+            auto len = XLENGTH(args[i]);
+            for (long j = 0; j < len; ++j) {
+                assert(pos < total);
+                // We handle LGL and INT in the same case here. That is fine,
+                // because they are essentially the same type.
+                SLOWASSERT(NA_INTEGER == NA_LOGICAL);
+                if (type == REALSXP) {
+                    if (TYPEOF(args[i]) == REALSXP) {
+                        REAL(res)[pos++] = REAL(args[i])[j];
+                    } else {
+                        if (INTEGER(args[i])[j] == NA_INTEGER) {
+                            REAL(res)[pos++] = NA_REAL;
+                        } else {
+                            REAL(res)[pos++] = INTEGER(args[i])[j];
+                        }
+                    }
+                } else {
+                    INTEGER(res)[pos++] = INTEGER(args[i])[j];
+                }
+            }
+        }
+        return res;
+    }
+
+    case 109: { // "vector"
+        if (nargs != 2)
+            return nullptr;
+        if (TYPEOF(args[0]) != STRSXP)
+            return nullptr;
+        if (XLENGTH(args[0]) != 1)
+            return nullptr;
+        size_t length = -1;
+        if (IS_SIMPLE_SCALAR(args[1], INTSXP)) {
+            if (INTEGER(args[1])[0] == NA_INTEGER)
+                return nullptr;
+            length = INTEGER(args[1])[0];
+        } else if (IS_SIMPLE_SCALAR(args[1], REALSXP)) {
+            if (REAL(args[1])[0] == NA_REAL)
+                return nullptr;
+            length = REAL(args[1])[0];
+        } else {
+            return nullptr;
+        }
+        int type = str2type(CHAR(args[0]));
+
+        switch (type) {
+        case LGLSXP:
+        case INTSXP: {
+            auto res = allocVector(type, length);
+            Memzero(INTEGER(res), length);
+            return res;
+        }
+        case CPLXSXP: {
+            auto res = allocVector(type, length);
+            Memzero(COMPLEX(res), length);
+            return res;
+        }
+        case RAWSXP: {
+            auto res = allocVector(type, length);
+            Memzero(RAW(res), length);
+            return res;
+        }
+        case REALSXP: {
+            auto res = allocVector(type, length);
+            Memzero(REAL(res), length);
+            return res;
+        }
+        case STRSXP:
+        case EXPRSXP:
+        case VECSXP:
+            return allocVector(type, length);
+        case LISTSXP:
+            if (length > INT_MAX)
+                return nullptr;
+            return allocList((int)length);
+        default:
+            return nullptr;
+        }
+        assert(false);
+    }
+
+    case 412: { // "list"
+        // "lists" at the R level are VECSXP's in the implementation
+        auto res = Rf_allocVector(VECSXP, nargs);
+        for (size_t i = 0; i < nargs; ++i)
+            SET_VECTOR_ELT(res, i, args[i]);
+        return res;
+    }
+
+    case 399: { // "is.vector"
+        if (nargs != 1)
+            return nullptr;
+
+        auto arg = args[0];
+        if (XLENGTH(arg) != 1)
+            return nullptr;
+        return TYPEOF(arg) == VECSXP ? R_TrueValue : R_FalseValue;
+    }
+
+    case 395: { // "is.na"
+        if (nargs != 1)
+            return nullptr;
+
+        auto arg = args[0];
+        if (XLENGTH(arg) != 1)
+            return nullptr;
+
+        switch (TYPEOF(arg)) {
+        case INTSXP:
+            return INTEGER(arg)[0] == NA_INTEGER ? R_TrueValue : R_FalseValue;
+        case LGLSXP:
+            return LOGICAL(arg)[0] == NA_LOGICAL ? R_TrueValue : R_FalseValue;
+        case REALSXP:
+            return ISNAN(REAL(arg)[0]) ? R_TrueValue : R_FalseValue;
+        default:
+            return nullptr;
+        }
+        assert(false);
+    }
+
+    case 393: { // "is.function"
+        if (nargs != 1)
+            return nullptr;
+        return isFunction(args[0]) ? R_TrueValue : R_FalseValue;
+    }
+
+    case 507: { // "islistfactor"
+        if (nargs != 2)
+            return nullptr;
+        auto n = XLENGTH(args[0]);
+        if (n == 0 || !isVectorList(args[0]))
+            return R_FalseValue;
+        int recursive = asLogical(args[1]);
+        if (recursive)
+            return nullptr;
+
+        for (int i = 0; i < n; i++)
+            if (!isFactor(VECTOR_ELT(args[0], i)))
+                return R_FalseValue;
+
+        return R_TrueValue;
+    }
+
+    case 88: { // "length"
+        if (nargs != 1)
+            return nullptr;
+
+        switch (TYPEOF(args[0])) {
+        case INTSXP:
+        case REALSXP:
+        case LGLSXP:
+            return Rf_ScalarInteger(XLENGTH(args[0]));
+        default:
+            return nullptr;
+        }
+        assert(false);
+    }
+
+    case 386: { // "is.numeric"
+        if (nargs != 1)
+            return nullptr;
+        return isNumeric(args[0]) && !isLogical(args[0]) ? R_TrueValue
+                                                         : R_FalseValue;
+    }
+
+    case 387: { // "is.matrix"
+        if (nargs != 1)
+            return nullptr;
+        return isMatrix(args[0]) ? R_TrueValue : R_FalseValue;
+    }
+
+    case 388: { // "is.array"
+        if (nargs != 1)
+            return nullptr;
+        return isArray(args[0]) ? R_TrueValue : R_FalseValue;
+    }
+
+    case 389: { // "is.atomic" (389)
+        if (nargs != 1)
+            return nullptr;
+        switch (TYPEOF(args[0])) {
+        case NILSXP:
+        /* NULL is atomic (S compatibly), but not in isVectorAtomic(.) */
+        case CHARSXP:
+        case LGLSXP:
+        case INTSXP:
+        case REALSXP:
+        case CPLXSXP:
+        case STRSXP:
+        case RAWSXP:
+            return R_TrueValue;
+        default:
+            return R_FalseValue;
+        }
+        assert(false);
+    }
+
+    case 384: { // "is.object" (384)
+        if (nargs != 1)
+            return nullptr;
+        return OBJECT(args[0]) ? R_TrueValue : R_FalseValue;
+    }
+    }
+
+    return nullptr;
+}
+}
